+++ scripts/test-coverage.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 set -o errexit
-set -o nounset
 set -o pipefail
 set -x
 
@@ -8,4 +7,4 @@ rm -rf coverage test-build
 
 tsc --project tsconfig.test.json
 
-c8 --include src --include test-build/src --reporter text --reporter lcov playwright test
+c8 --include src --include test-build/src --reporter text --reporter lcov playwright test $1 $2 $3