+++ test/escapeHTML.test.ts
@@ -1,5 +1,12 @@
-import { test } from '@playwright/test';
+import { expect, test } from '@playwright/test';
+import { escapeHTML } from '../src/exdom.js';
 
 test('escapeHTML', () => {
-  // TODO
+  expect(escapeHTML('')).toEqual('');
+  expect(escapeHTML('test')).toEqual('test');
+  expect(escapeHTML('<script>')).toEqual('&lt;script&gt;');
+  expect(escapeHTML('"hello"')).toEqual('&quot;hello&quot;');
+  expect(escapeHTML('<a href="hackerz">&amp;</a>')).toEqual(
+    '&lt;a href=&quot;hackerz&quot;&gt;&amp;amp;&lt;/a&gt;',
+  );
 });